fix(tasks): resolve all technical debt from dedup refactor
- Fix 4 pre-existing build errors (unused import, missing type imports,
  exactOptionalPropertyTypes incompatibility)
- Remove redundant SendProgressDeps type, unify to TaskNotifierDeps
- Fix toErrorMessage: Error objects now use String() fallback for empty
  messages instead of JSON serialization producing '{}'
- Rename stuck-detector.test → task-monitor.test to match source
- Merge progress-tracker + progress-analyzer tests into progress.test
- Update AGENTS.md architecture diagram and directory listing
